feat: Add docker-compose file for MongoDB and Redis containers
This commit adds a docker-compose file to the server folder that can be used to run MongoDB and Redis containers with a single command. The file uses the official MongoDB and Redis images and maps the container's default ports to the host. The containers are set to run in the foreground for ease of use and debugging.